adding final refinements following comments and discusssion
